+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changelog
 * Rename ``supported`` property of managers to ``fresh``.
 * Allow restriction of package managers to a platform. Closes #7.
 * Include ``supported`` property in ``mpm managers`` sub-command.
-* Add optional sub-menu rendering in BitBar plugin. Closes #23.
+* Add optional submenu rendering for BitBar plugin. Closes #23.
 * Move ``Upgrade all`` menu entry to the bottom of earch section in BitBar
   plugin.
 * Allow destructive unittests in Travis CI jobs.